requirements.txt
accelerate
bitsandbytes
cairosvg
captum
cython
datasets
einops
evaluate
ipykernel
kaleido
matplotlib
open_clip_torch
optuna
peft
pillow
plotly
protobuf
qwen-vl-utils
scikit-learn
scikit-image
torch
torchvision
transformers==4.49.0
wandb